Integration with backend and services

+++ api/api/controllers/user-controller.js
@@ -9,9 +9,9 @@ const Stock = require('../models/stock-model').model
 
 // CREATE A NEW USER
 const create = async (request, response) => {
+  let createdUser
   try {
     const dbUser = await User.findOne({ username: request.body.username.toLowerCase() })
-
     if (!dbUser) {
       const { username, password } = request.body
 
@@ -23,7 +23,7 @@ const create = async (request, response) => {
         password: hash
       })
 
-      const createdUser = await user.save()
+      createdUser = await user.save()
 
       const operation = await axios.post(`${process.env.API_URL}/accounts/${user._id}`)
 
@@ -48,7 +48,9 @@ const create = async (request, response) => {
       status: error.response.status || 500,
       message: error.response.data.message || 'Operation not completed'
     }
-    await User.findByIdAndDelete(createdUser._id)
+    if (createdUser !== undefined) {
+      await User.findByIdAndDelete(createdUser._id)
+    }
     response.status(err.status).json({ message: err.message })
   }
 }
@@ -78,9 +80,10 @@ const read = async (request, response) => {
 
 // UPDATE AN USER
 const update = async (request, response) => {
+  console.log(request.headers)
   try {
     const { stock, action } = request.body
-    const id = mongoose.Types.ObjectId(request.params.id) 
+    const id = mongoose.Types.ObjectId(request.params.id)
 
     const dbUser = await User.findById(id)
 
@@ -90,56 +93,56 @@ const update = async (request, response) => {
 
     let index
 
-    switch(action){
+    switch (action) {
       case 'BUY':
-        index = dbUser.stocks.map((userStock) => { return userStock.name; }).indexOf(stock.name);
-        if(index === -1){
-          let bought = new Stock({
+        index = dbUser.stocks.map((userStock) => { return userStock.name }).indexOf(stock.name)
+        if (index === -1) {
+          const bought = new Stock({
             _id: mongoose.Types.ObjectId(),
             name: stock.name,
             quantity: stock.quantity
           })
           dbUser.stocks.push(bought)
         } else {
-          dbUser.stocks[index]['quantity'] += stock.quantity
+          dbUser.stocks[index].quantity += stock.quantity
         }
-        break;
+        break
       case 'SELL':
-        index = dbUser.stocks.map((userStock) => { return userStock.name; }).indexOf(stock.name);
-          if(index === -1){
-            return response.status(404).json({ error: true, message: 'User does not have this stock on the account'})
+        index = dbUser.stocks.map((userStock) => { return userStock.name }).indexOf(stock.name)
+        if (index === -1) {
+          return response.status(404).json({ error: true, message: 'User does not have this stock on the account' })
+        } else {
+          if (dbUser.stocks[index].quantity - stock.quantity < 0) {
+            return response.status(400).json({ error: true, message: `User doest not have ${stock.quantity} shares of this stock to sell` })
           } else {
-            if(dbUser.stocks[index]['quantity'] - stock.quantity < 0){
-              return response.status(400).json({ error: true, message: `User doest not have ${stock.quantity} shares of this stock to sell` })
-            } else {
-              dbUser.stocks[index]['quantity'] -= stock.quantity
-              if(dbUser.stocks[index]['quantity'] == 0){
-                dbUser.stocks.splice(index, 1)
-              }
+            dbUser.stocks[index].quantity -= stock.quantity
+            if (dbUser.stocks[index].quantity === 0) {
+              dbUser.stocks.splice(index, 1)
             }
           }
-        break;
-      
-      default:
-        return response.status(400).json({ error: true, message: 'Action not registered'})
-      }
-      
-      await axios.patch(`${process.env.API_URL}/accounts/${dbUser._id}`, { action, value: stock.value })
-      
-      await dbUser.save()
+        }
+        break
 
-      response.status(200).json({ message: 'Operation successfull' })
+      default:
+        return response.status(400).json({ error: true, message: 'Action not registered' })
+    }
+
+    await axios.patch(`${process.env.API_URL}/accounts/${dbUser._id}`, { action, value: stock.value })
+
+    await dbUser.save()
+
+    response.status(200).json({ message: 'Operation successfull' })
   } catch (error) {
     const err = {
       status: error.response.status || 500,
       message: error.response.data.message || 'Operation not completed'
     }
-    response.status(err.status).json({ message: err.message})
+    response.status(err.status).json({ message: err.message })
   }
 }
 
 module.exports = {
   create,
   read,
-  update,
+  update
 }
